Payroll: Unify test files
