+++ future-apps/payroll/test/helpers/numbers.js
@@ -1,9 +1,6 @@
-const SECONDS_IN_A_YEAR = 31557600 // 365.25 days
-
 module.exports = web3 => {
   const bn = x => new web3.BigNumber(x)
   const bigExp = (x, y) => bn(x).mul(bn(10).pow(bn(y)))
-  const salary = (x, y) => bigExp(x, y).dividedToIntegerBy(SECONDS_IN_A_YEAR)
 
   const maxUint = (e) => bn(2).pow(bn(e)).sub(bn(1))
   const maxUint64 = () => maxUint(64)
@@ -12,7 +9,6 @@ module.exports = web3 => {
   return {
     bn,
     bigExp,
-    salary,
     maxUint64,
     maxUint256,
   }
