--- future-apps/payroll/test/payroll_initialize.js
@@ -1,85 +1,0 @@
-const { assertRevert } = require('@aragon/test-helpers/assertThrow')
-
-const ACL = artifacts.require('ACL')
-const Payroll = artifacts.require('PayrollMock')
-const PriceFeed = artifacts.require('PriceFeedMock')
-
-const getEvent = (receipt, event) => receipt.logs.find(l => l.event === event).args
-const getEventArgument = (receipt, event, arg) => getEvent(receipt, event)[arg]
-
-const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
-
-contract('Payroll, initialization,', function(accounts) {
-  const [owner] = accounts
-  const { deployErc20TokenAndDeposit, redistributeEth, getDaoFinanceVault } = require('./helpers.js')(owner)
-
-  const USD_DECIMALS= 18
-  const RATE_EXPIRATION_TIME = 1000
-
-  let dao, acl, payroll, payrollBase, finance, vault, priceFeed, usdToken
-
-  before('setup base apps and tokens', async () => {
-    const daoFinanceVault = await getDaoFinanceVault()
-    dao = daoFinanceVault.dao
-    finance = daoFinanceVault.finance
-    vault = daoFinanceVault.vault
-    acl = ACL.at(await dao.acl())
-
-    priceFeed = await PriceFeed.new()
-    payrollBase = await Payroll.new()
-    usdToken = await deployErc20TokenAndDeposit(owner, finance, vault, 'USD', USD_DECIMALS)
-
-    await redistributeEth(accounts, finance)
-  })
-
-  beforeEach('create payroll instance', async () => {
-    const receipt = await dao.newAppInstance('0x4321', payrollBase.address, '0x', false, { from: owner })
-    payroll = Payroll.at(getEventArgument(receipt, 'NewAppProxy', 'proxy'))
-  })
-
-  describe('initialize', function () {
-    const from = owner
-
-    it('cannot initialize the base app', async () => {
-      assert(await payrollBase.isPetrified(), 'base payroll app should be petrified')
-      await assertRevert(payrollBase.initialize(finance.address, usdToken.address, priceFeed.address, RATE_EXPIRATION_TIME, { from }), 'INIT_ALREADY_INITIALIZED')
-    })
-
-    context('when it has not been initialized yet', function () {
-      it('can be initialized with a zero denomination token', async () => {
-        payroll.initialize(finance.address, ZERO_ADDRESS, priceFeed.address, RATE_EXPIRATION_TIME, { from })
-        assert.equal(await payroll.denominationToken(), ZERO_ADDRESS, 'denomination token does not match')
-      })
-
-      it('reverts when passing an expiration time lower than or equal to a minute', async () => {
-        const ONE_MINUTE = 60
-        await assertRevert(payroll.initialize(finance.address, usdToken.address, priceFeed.address, ONE_MINUTE, { from }), 'PAYROLL_EXPIRY_TIME_TOO_SHORT')
-      })
-
-      it('reverts when passing an invalid finance instance', async () => {
-        await assertRevert(payroll.initialize(ZERO_ADDRESS, usdToken.address, priceFeed.address, RATE_EXPIRATION_TIME, { from }), 'PAYROLL_FINANCE_NOT_CONTRACT')
-      })
-
-      it('reverts when passing an invalid feed instance', async () => {
-        await assertRevert(payroll.initialize(finance.address, usdToken.address, ZERO_ADDRESS, RATE_EXPIRATION_TIME, { from }), 'PAYROLL_FEED_NOT_CONTRACT')
-      })
-    })
-
-    context('when it has already been initialized', function () {
-      beforeEach('initialize payroll app', async () => {
-        await payroll.initialize(finance.address, usdToken.address, priceFeed.address, RATE_EXPIRATION_TIME, { from })
-      })
-
-      it('cannot be initialized again', async () => {
-        await assertRevert(payroll.initialize(finance.address, usdToken.address, priceFeed.address, RATE_EXPIRATION_TIME, { from }), 'INIT_ALREADY_INITIALIZED')
-      })
-
-      it('has a price feed instance, a finance instance, a denomination token and a rate expiration time', async () => {
-        assert.equal(await payroll.feed(), priceFeed.address, 'feed address does not match')
-        assert.equal(await payroll.finance(), finance.address, 'finance address should match')
-        assert.equal(await payroll.denominationToken(), usdToken.address, 'denomination token does not match')
-        assert.equal(await payroll.rateExpiryTime(), RATE_EXPIRATION_TIME, 'rate expiration time does not match')
-      })
-    })
-  })
-})
